requirements.txt
openai==1.59.3
pydub==0.25.1
python-dotenv==1.0.1
